backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
pymysql>=1.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
langgraph>=0.2.0
zai-sdk
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
